+++ app/src/main/java/com/example/arsalan/mygym/viewModels/TrainerWorkoutPlanReqVm.java
@@ -1,15 +1,11 @@
 package com.example.arsalan.mygym.viewModels;
 
 import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
-import androidx.lifecycle.ViewModelProviders;
 
-import com.example.arsalan.mygym.models.Token;
 import com.example.arsalan.mygym.models.WorkoutPlanReq;
 import com.example.arsalan.mygym.repository.TrainerWorkoutPlanReqRepository;
 
-import java.security.PublicKey;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -28,7 +24,7 @@ public class TrainerWorkoutPlanReqVm extends ViewModel {
     public void init(long trainerId){
         mWorkoutPlanListLv=  mRepository.getListLive(trainerId);
     }
-    public LiveData<Integer> cancelRequest(int requestId){
+    public LiveData<Integer> cancelRequest(long requestId){
         return mRepository.cancelWorkoutRequest(requestId);
     }
